requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mkl-fft==1.0.2
mkl-random==1.0.1
numpy==1.14.5
python-dotenv==0.8.2
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
